+++ src/main/java/net/client/handler/WsClientHandler.java
@@ -162,7 +162,7 @@ public class WsClientHandler extends SimpleChannelInboundHandler<WebSocketFrame>
 	}
 
 	@Override
-	public void sendMessage(int clientId, int msgId, int mapId, Message msg, int sequence) {
+	public void sendMessage(int clientId, int msgId, long mapId, Message msg, int sequence) {
 		channel.writeAndFlush(maker.wrap(clientId, msgId, mapId, msg, sequence));
 	}
