+++ src/main/java/net/connect/handle/ConnectHandler.java
@@ -269,7 +269,7 @@ public class ConnectHandler extends ChannelInboundHandlerAdapter implements Send
 	}
 
 	@Override
-	public void sendMessage(int clientId, int msgId, int mapId, Message msg, int sequence) {
+	public void sendMessage(int clientId, int msgId, long mapId, Message msg, int sequence) {
 		channel.writeAndFlush(maker.wrap(clientId, msgId, mapId, msg, sequence));
 	}
